Adding tasks for forcing profile states
These tasks are really intended for demo purposes only. By running the
tasks you can make sure all of your users are in the appropriate
state. This does not, however, mean that those states will be valid
in the larger sense (i.e. we don't really authenticate each users
connection with Orcid, but are instead spoofing that for our system).